Fix Quote

+++ src/main/java/com/zetalabs/indumelec/model/QuoteDetail.java
@@ -3,6 +3,7 @@ package com.zetalabs.indumelec.model;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @Entity
@@ -23,8 +24,8 @@ public class QuoteDetail {
     private String measure;
 
     @Column(name = "quantity")
-    private Integer quantity;
+    private BigDecimal quantity;
 
     @Column(name = "price")
-    private Double price;
+    private BigDecimal price;
 }
